Safari-609.1.11.1/LayoutTests/http/tests/contentdispositionattachmentsandbox/at-import-stylesheets-disabled-expected.txt

CONSOLE MESSAGE: line 4: Unsafe attempt to load URL data:text/css,body::after { content: 'FAIL'; } from document with Content-Disposition: attachment at URL http://127.0.0.1:8000/contentdispositionattachmentsandbox/resources/at-import-stylesheets-frame.php.
This test verifies that @imported stylesheets are disabled when 'Content-Disposition: attachment' sandboxing is enabled. A security error will be logged to the console if the test passes.
